+++ HISTORY.rst
@@ -21,6 +21,7 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* Create date ranges (accounting periods) for 2017
 * Install module sf_terms_and_conditions
 * RMA decision fields adaptation